moved Ingredient & IngredientItem to shared folder
