+++ frontend/foodieland-app/components/shared/IngredientItem.tsx
